+++ core/embed/Cargo.toml
@@ -4,6 +4,7 @@ cargo-features = ["panic-immediate-abort"]
 members = [
     "models",
     "rtl",
+    "crypto",
     "sys",
     "sec",
     "io",
@@ -93,6 +94,7 @@ zeroize = { version = "1.8.2", default-features = false, features = ["zeroize_de
 
 models = { path = "models" }
 rtl = { path = "rtl" }
+crypto = { path = "crypto" }
 sys = { path = "sys" }
 sec = { path = "sec" }
 io = { path = "io" }
@@ -113,7 +115,7 @@ hex = "0.4.3"
 owo-colors = "4.3.0"
 pathdiff = "0.2.3"
 pkg-config = "0.3.28"
-rand = { version = "0.10.1", default-features = false }
+rand = { version = "0.10.0", default-features = false }
 serde = "1.0.228"
 serde_json = "1.0.145"
 tempfile = { version = "3.25.0", default-features = false }
