+++ core/embed/crypto/src/aesgcm.rs
@@ -1,9 +1,10 @@
 use core::pin::Pin;
 
+use rtl::error::ensure;
 use zeroize::Zeroize;
 
 use super::memory::Memory;
-use super::{consteq, ffi, Error};
+use super::{Error, consteq, ffi};
 
 // Tag size is a parameter but we fix it to 16 here for simplicity.
 pub const TAG_SIZE: usize = 16;
@@ -276,61 +277,61 @@ mod test {
             ciphertext: "42831ec2217774244b7221b784d0d49ce3aa212f2c02a4e035c17e2329aca12e21d514b25466931c7d8f6a5aac84aa051ba30b396a0aac973d58e091473f5985",
             tag: "4d5c2af327cd64a62cf35abd2ba6fab4",
         },
-		Vector {
-			key: "feffe9928665731c6d6a8f9467308308",
-			iv: "cafebabefacedbaddecaf888",
-			aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
+        Vector {
+            key: "feffe9928665731c6d6a8f9467308308",
+            iv: "cafebabefacedbaddecaf888",
+            aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
             plaintext: "d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b39",
-			ciphertext: "42831ec2217774244b7221b784d0d49ce3aa212f2c02a4e035c17e2329aca12e21d514b25466931c7d8f6a5aac84aa051ba30b396a0aac973d58e091",
-			tag: "5bc94fbc3221a5db94fae95ae7121a47",
+            ciphertext: "42831ec2217774244b7221b784d0d49ce3aa212f2c02a4e035c17e2329aca12e21d514b25466931c7d8f6a5aac84aa051ba30b396a0aac973d58e091",
+            tag: "5bc94fbc3221a5db94fae95ae7121a47",
         },
-		Vector {
-			key: "feffe9928665731c6d6a8f9467308308",
-			iv: "cafebabefacedbad",
-			aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
+        Vector {
+            key: "feffe9928665731c6d6a8f9467308308",
+            iv: "cafebabefacedbad",
+            aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
             plaintext: "d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b39",
-			ciphertext: "61353b4c2806934a777ff51fa22a4755699b2a714fcdc6f83766e5f97b6c742373806900e49f24b22b097544d4896b424989b5e1ebac0f07c23f4598",
-			tag: "3612d2e79e3b0785561be14aaca2fccb",
+            ciphertext: "61353b4c2806934a777ff51fa22a4755699b2a714fcdc6f83766e5f97b6c742373806900e49f24b22b097544d4896b424989b5e1ebac0f07c23f4598",
+            tag: "3612d2e79e3b0785561be14aaca2fccb",
         },
-		Vector {
-			key: "feffe9928665731c6d6a8f9467308308",
-			iv: "9313225df88406e555909c5aff5269aa6a7a9538534f7da1e4c303d2a318a728c3c0c95156809539fcf0e2429a6b525416aedbf5a0de6a57a637b39b",
-			aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
+        Vector {
+            key: "feffe9928665731c6d6a8f9467308308",
+            iv: "9313225df88406e555909c5aff5269aa6a7a9538534f7da1e4c303d2a318a728c3c0c95156809539fcf0e2429a6b525416aedbf5a0de6a57a637b39b",
+            aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
             plaintext: "d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b39",
-			ciphertext: "8ce24998625615b603a033aca13fb894be9112a5c3a211a8ba262a3cca7e2ca701e4a9a4fba43c90ccdcb281d48c7c6fd62875d2aca417034c34aee5",
-			tag: "619cc5aefffe0bfa462af43c1699d050",
+            ciphertext: "8ce24998625615b603a033aca13fb894be9112a5c3a211a8ba262a3cca7e2ca701e4a9a4fba43c90ccdcb281d48c7c6fd62875d2aca417034c34aee5",
+            tag: "619cc5aefffe0bfa462af43c1699d050",
         },
-		Vector {
-			key: "000000000000000000000000000000000000000000000000",
-			iv: "000000000000000000000000",
+        Vector {
+            key: "000000000000000000000000000000000000000000000000",
+            iv: "000000000000000000000000",
             aad: "",
             plaintext: "",
             ciphertext: "",
-			tag: "cd33b28ac773f74ba00ed1f312572435",
+            tag: "cd33b28ac773f74ba00ed1f312572435",
         },
-		Vector {
-			key: "000000000000000000000000000000000000000000000000",
-			iv: "000000000000000000000000",
+        Vector {
+            key: "000000000000000000000000000000000000000000000000",
+            iv: "000000000000000000000000",
             aad: "",
             plaintext: "00000000000000000000000000000000",
-			ciphertext: "98e7247c07f0fe411c267e4384b0f600",
-			tag: "2ff58d80033927ab8ef4d4587514f0fb",
+            ciphertext: "98e7247c07f0fe411c267e4384b0f600",
+            tag: "2ff58d80033927ab8ef4d4587514f0fb",
         },
-		Vector {
-			key: "feffe9928665731c6d6a8f9467308308feffe9928665731c",
-			iv: "cafebabefacedbaddecaf888",
+        Vector {
+            key: "feffe9928665731c6d6a8f9467308308feffe9928665731c",
+            iv: "cafebabefacedbaddecaf888",
             aad: "",
             plaintext: "d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b391aafd255",
-			ciphertext: "3980ca0b3c00e841eb06fac4872a2757859e1ceaa6efd984628593b40ca1e19c7d773d00c144c525ac619d18c84a3f4718e2448b2fe324d9ccda2710acade256",
-			tag: "9924a7c8587336bfb118024db8674a14",
+            ciphertext: "3980ca0b3c00e841eb06fac4872a2757859e1ceaa6efd984628593b40ca1e19c7d773d00c144c525ac619d18c84a3f4718e2448b2fe324d9ccda2710acade256",
+            tag: "9924a7c8587336bfb118024db8674a14",
         },
-		Vector {
-			key: "feffe9928665731c6d6a8f9467308308feffe9928665731c",
-			iv: "cafebabefacedbaddecaf888",
-			aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
+        Vector {
+            key: "feffe9928665731c6d6a8f9467308308feffe9928665731c",
+            iv: "cafebabefacedbaddecaf888",
+            aad: "feedfacedeadbeeffeedfacedeadbeefabaddad2",
             plaintext: "d9313225f88406e5a55909c5aff5269a86a7a9531534f7da2e4c303d8a318a721c3c0c95956809532fcf0e2449a6b525b16aedf5aa0de657ba637b39",
-			ciphertext: "3980ca0b3c00e841eb06fac4872a2757859e1ceaa6efd984628593b40ca1e19c7d773d00c144c525ac619d18c84a3f4718e2448b2fe324d9ccda2710",
-			tag: "2519498e80f1478f37ba55bd6d27618c",
+            ciphertext: "3980ca0b3c00e841eb06fac4872a2757859e1ceaa6efd984628593b40ca1e19c7d773d00c144c525ac619d18c84a3f4718e2448b2fe324d9ccda2710",
+            tag: "2519498e80f1478f37ba55bd6d27618c",
         },
         // test vectors from test_trezor.wire.thp.crypto.py
         Vector {
