+++ core/embed/crypto/src/ffi.rs
@@ -1,3 +1,4 @@
 #![allow(non_camel_case_types)]
+#![allow(dead_code)]
 
 include!(concat!(env!("OUT_DIR"), "/crypto.rs"));
