+++ core/embed/crypto/src/hmac.rs
@@ -3,7 +3,7 @@ use core::pin::Pin;
 use zeroize::{Zeroize, ZeroizeOnDrop};
 
 use super::ffi;
-use super::memory::{init_ctx, Memory};
+use super::memory::{Memory, init_ctx};
 
 pub const DIGEST_SIZE: usize = ffi::SHA256_DIGEST_LENGTH as usize;
 pub type Digest = [u8; DIGEST_SIZE];
@@ -51,43 +51,42 @@ pub fn digest(key: &[u8], data: &[u8]) -> Digest {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::strutil::hexlify;
 
-    const HMAC_SHA256_EMPTY: &[u8] =
-        b"b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad";
+    const HMAC_SHA256_EMPTY: &str =
+        "b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad";
     // RFC 4231
-    const HMAC_SHA256_VECTORS: &[(&[u8], &[u8], &[u8])] = &[
+    const HMAC_SHA256_VECTORS: &[(&[u8], &[u8], &str)] = &[
         (
             &[0x0b; 20],
             b"Hi There",
-            b"b0344c61d8db38535ca8afceaf0bf12b881dc200c9833da726e9376c2e32cff7",
+            "b0344c61d8db38535ca8afceaf0bf12b881dc200c9833da726e9376c2e32cff7",
         ),
         (
             b"Jefe",
             b"what do ya want for nothing?",
-            b"5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843",
+            "5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843",
         ),
 
         (
             &[0xaa; 20],
             &[0xdd; 50],
-            b"773ea91e36800e46854db8ebd09181a72959098b3ef8c122d9635514ced565fe",
+            "773ea91e36800e46854db8ebd09181a72959098b3ef8c122d9635514ced565fe",
         ),
         (
             &[0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19],
             &[0xcd; 50],
-            b"82558a389a443c0ea4cc819899f2083a85f0faa3e578f8077a2e3ff46729665b",
+            "82558a389a443c0ea4cc819899f2083a85f0faa3e578f8077a2e3ff46729665b",
         ),
         // skipping case with truncation
         (
             &[0xaa; 131],
             b"Test Using Larger Than Block-Size Key - Hash Key First",
-            b"60e431591ee0b67f0d8a26aacbf5b77f8e0bc6213728c5140546040f0ee37f54",
+            "60e431591ee0b67f0d8a26aacbf5b77f8e0bc6213728c5140546040f0ee37f54",
         ),
         (
             &[0xaa; 131],
             b"This is a test using a larger than block-size key and a larger than block-size data. The key needs to be hashed before being used by the HMAC algorithm.",
-            b"9b09ffa71b942fcb27635fbcd5b0e944bfdc63644f0713938a7f51535c3a35e2",
+            "9b09ffa71b942fcb27635fbcd5b0e944bfdc63644f0713938a7f51535c3a35e2",
         ),
         (
             b"",
@@ -96,22 +95,17 @@ mod test {
         ),
     ];
 
-    fn hexdigest(key: &[u8], data: &[u8]) -> [u8; DIGEST_SIZE * 2] {
-        let mut out_hex = [0u8; DIGEST_SIZE * 2];
-
-        let digest = digest(key, data);
-        hexlify(&digest, &mut out_hex);
-        out_hex
+    fn hexdigest(key: &[u8], data: &[u8]) -> String {
+        hex::encode(digest(key, data))
     }
 
     #[test]
     fn test_empty_ctx() {
         let mut out = [0u8; DIGEST_SIZE];
-        let mut out_hex = [0u8; DIGEST_SIZE * 2];
 
         init_ctx!(HmacSha256, ctx, b"");
         ctx.finalize_into(&mut out);
-        hexlify(&out, &mut out_hex);
+        let out_hex = hex::encode(out);
 
         assert_eq!(out_hex, HMAC_SHA256_EMPTY);
     }
