+++ core/embed/crypto/src/merkle.rs
