+++ core/embed/crypto/src/sha512.rs
@@ -3,7 +3,7 @@ use core::pin::Pin;
 use zeroize::{Zeroize, ZeroizeOnDrop};
 
 use super::ffi;
-use super::memory::{init_ctx, Memory};
+use super::memory::{Memory, init_ctx};
 
 pub const BLOCK_SIZE: usize = ffi::SHA512_BLOCK_LENGTH as usize;
 pub const DIGEST_SIZE: usize = ffi::SHA512_DIGEST_LENGTH as usize;
@@ -53,42 +53,36 @@ pub fn digest(data: &[u8]) -> Digest {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::strutil::hexlify;
 
-    const SHA512_EMPTY: &[u8] = b"cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e";
-    const SHA512_VECTORS: &[(&[u8], &[u8])] = &[
+    const SHA512_EMPTY: &str = "cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e";
+    const SHA512_VECTORS: &[(&[u8], &str)] = &[
         (b"", SHA512_EMPTY),
         (
             b"abc",
-            b"ddaf35a193617abacc417349ae20413112e6fa4e89a97ea20a9eeee64b55d39a2192992a274fc1a836ba3c23a3feebbd454d4423643ce80e2a9ac94fa54ca49f",
+            "ddaf35a193617abacc417349ae20413112e6fa4e89a97ea20a9eeee64b55d39a2192992a274fc1a836ba3c23a3feebbd454d4423643ce80e2a9ac94fa54ca49f",
         ),
         (
             b"abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq",
-            b"204a8fc6dda82f0a0ced7beb8e08a41657c16ef468b228a8279be331a703c33596fd15c13b1b07f9aa1d3bea57789ca031ad85c7a71dd70354ec631238ca3445",
+            "204a8fc6dda82f0a0ced7beb8e08a41657c16ef468b228a8279be331a703c33596fd15c13b1b07f9aa1d3bea57789ca031ad85c7a71dd70354ec631238ca3445",
         ),
         (
             b"abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmnhijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu",
-            b"8e959b75dae313da8cf4f72814fc143f8f7779c6eb9f7fa17299aeadb6889018501d289e4900f7e4331b99dec4b5433ac7d329eeb6dd26545e96e55b874be909",
+            "8e959b75dae313da8cf4f72814fc143f8f7779c6eb9f7fa17299aeadb6889018501d289e4900f7e4331b99dec4b5433ac7d329eeb6dd26545e96e55b874be909",
         ),
     ];
 
-    fn hexdigest(data: &[u8]) -> [u8; DIGEST_SIZE * 2] {
-        let mut out_hex = [0u8; DIGEST_SIZE * 2];
-
-        let digest = digest(data);
-        hexlify(&digest, &mut out_hex);
-        out_hex
+    fn hexdigest(data: &[u8]) -> String {
+        hex::encode(digest(data))
     }
 
     #[test]
     fn test_empty_ctx() {
         let mut out = [0u8; DIGEST_SIZE];
-        let mut out_hex = [0u8; DIGEST_SIZE * 2];
 
         init_ctx!(Sha512, ctx);
         ctx.finalize_into(&mut out);
-        hexlify(&out, &mut out_hex);
 
+        let out_hex = hex::encode(out);
         assert_eq!(out_hex, SHA512_EMPTY);
     }
