+++ core/embed/projects/bootloader/Cargo.toml
@@ -13,6 +13,7 @@ trezor_lib.workspace = true
 io.workspace = true
 sec.workspace = true
 sys.workspace = true
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 
@@ -47,7 +48,7 @@ default = [
     "io/secure_mode",
     "io/usb_iface_wire",
     "models/bootloader",
-    "rtl/ed25519_no_precomp",
+    "crypto/ed25519_no_precomp",
 
     "trezor_lib/with_new_crates", # only for transitional period
