+++ core/embed/projects/firmware/Cargo.toml
@@ -13,6 +13,7 @@ trezor_lib.workspace = true
 io.workspace = true
 sec.workspace = true
 sys.workspace = true
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 upymod.workspace = true
@@ -25,7 +26,7 @@ upymod.workspace = true
 
 block_on_vcp = ["sys/block_on_vcp"]
 bootloader_devel = ["sec/bootloader_devel"]
-benchmark = ["upymod/benchmark", "rtl/aes_gcm"]
+benchmark = ["upymod/benchmark", "crypto/aes_gcm"]
 clippy = ["trezor_lib/clippy"]
 dbg_console = ["sys/dbg_console", "trezor_lib/dbg_console"]
 debug = ["trezor_lib/debug"]
@@ -56,7 +57,7 @@ default = [
     "io/display",
     "io/fancy_fatal_error",
     "io/usb_iface_wire",
-    "rtl/secp256k1_zkp",
+    "crypto/secp256k1_zkp",
 
     "trezor_lib/with_new_crates", # only for transitional period
     "trezor_lib/with_upymod",     # only for transitional period
@@ -110,7 +111,7 @@ secret = ["sec/secret"]
 secure_aes = ["sec/secure_aes"]
 serial_number = ["upymod/serial_number", "trezor_lib/serial_number"]
 suspend = ["io/suspend"]
-thp = ["upymod/thp", "rtl/aes_gcm", "trezor_lib/thp"]
+thp = ["upymod/thp", "crypto/aes_gcm", "trezor_lib/thp"]
 tamper = ["sec/tamper"]
 telemetry = ["sec/telemetry", "upymod/telemetry", "trezor_lib/telemetry"]
 touch = ["io/touch", "upymod/touch", "trezor_lib/touch"]
@@ -119,8 +120,8 @@ tropic = ["sec/tropic", "upymod/tropic", "trezor_lib/tropic"]
 
 # Legacy coin support (T2T1 only)
 decred = ["upymod/decred"]
-eos = ["rtl/eos", "upymod/eos"]
-nem = ["rtl/nem", "upymod/nem"]
+eos = ["crypto/eos", "upymod/eos"]
+nem = ["crypto/nem", "upymod/nem"]
 
 display_mono = ["trezor_lib/display_mono"]
 display_rgb565 = ["trezor_lib/display_rgb565"]
