+++ core/embed/projects/prodtest/Cargo.toml
@@ -12,6 +12,7 @@ trezor_lib.workspace = true
 io.workspace = true
 sec.workspace = true
 sys.workspace = true
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 
@@ -45,9 +46,9 @@ default = [
     "io/usb_iface_vcp",
     "models/prodtest",
     "rtl/sprintf",
-    "rtl/mldsa",
-    "rtl/aes_gcm",
-    "rtl/noise",
+    "crypto/aes_gcm",
+    "crypto/mldsa",
+    "crypto/noise",
 
     "trezor_lib/with_new_crates", # only for transitional period
