+++ core/embed/projects/unix/Cargo.toml
@@ -13,6 +13,7 @@ trezor_lib.workspace = true
 io.workspace = true
 sec.workspace = true
 sys.workspace = true
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 upymod.workspace = true
@@ -24,7 +25,7 @@ upymod.workspace = true
 # --------------------------------------------------------------------------
 
 asan = ["models/asan"]
-benchmark = ["upymod/benchmark", "rtl/aes_gcm"]
+benchmark = ["upymod/benchmark", "crypto/aes_gcm"]
 clippy = ["trezor_lib/clippy"]
 dbg_console = ["sys/dbg_console", "trezor_lib/dbg_console"]
 debug = ["trezor_lib/debug"]
@@ -56,8 +57,8 @@ default = [
     "io/secure_mode",
     "io/usb_iface_wire",
     "sec/rdi",
-    "rtl/secp256k1_zkp",
-    "rtl/aes_gcm",
+    "crypto/aes_gcm",
+    "crypto/secp256k1_zkp",
 
     "trezor_lib/with_new_crates", # only for transitional period
     "trezor_lib/with_upymod", # only for transitional period
@@ -109,15 +110,15 @@ serial_number = ["upymod/serial_number", "trezor_lib/serial_number"]
 suspend = ["io/suspend"]
 tamper = ["sec/tamper"]
 telemetry = ["sec/telemetry", "upymod/telemetry", "trezor_lib/telemetry"]
-thp = ["upymod/thp", "rtl/aes_gcm", "trezor_lib/thp"]
+thp = ["upymod/thp", "crypto/aes_gcm", "trezor_lib/thp"]
 touch = ["io/touch", "upymod/touch", "trezor_lib/touch"]
 touch_wakeup = ["io/touch_wakeup", "upymod/touch_wakeup", "trezor_lib/touch_wakeup"]
 tropic = ["sec/tropic", "upymod/tropic", "trezor_lib/tropic"]
 
 # Legacy coin support
 decred = ["upymod/decred"]
-eos = ["rtl/eos", "upymod/eos"]
-nem = ["rtl/nem", "upymod/nem"]
+eos = ["crypto/eos", "upymod/eos"]
+nem = ["crypto/nem", "upymod/nem"]
 
 display_mono = ["trezor_lib/display_mono"]
 display_rgb565 = ["trezor_lib/display_rgb565"]
