+++ core/embed/rust/build.rs
@@ -16,8 +16,6 @@ fn main() {
     #[cfg(feature = "micropython")]
     generate_micropython_bindings();
     generate_trezorhal_bindings();
-    #[cfg(feature = "crypto")]
-    generate_crypto_bindings();
     #[cfg(all(feature = "test", not(feature = "with_new_crates")))]
     link_core_objects();
 
@@ -615,66 +613,6 @@ fn generate_trezorhal_bindings() {
         .generate()
         .expect("Unable to generate bindings")
         .write_to_file(PathBuf::from(out_dir).join("trezorhal.rs"))
-        .unwrap();
-}
-
-#[cfg(feature = "crypto")]
-fn generate_crypto_bindings() {
-    let out_dir = env::var("OUT_DIR").unwrap();
-
-    // Tell cargo to invalidate the built crate whenever the header changes.
-    println!("cargo:rerun-if-changed=crypto.h");
-
-    let bindings = prepare_bindings()
-        .header("crypto.h")
-        // aesgcm
-        .allowlist_type("gcm_ctx")
-        .no_copy("gcm_ctx")
-        .allowlist_function("gcm_init_and_key")
-        .allowlist_function("gcm_init_message")
-        .allowlist_function("gcm_encrypt")
-        .allowlist_function("gcm_decrypt")
-        .allowlist_function("gcm_auth_header")
-        .allowlist_function("gcm_compute_tag")
-        // curve25519
-        .allowlist_function("curve25519_scalarmult")
-        .allowlist_function("curve25519_scalarmult_basepoint")
-        // ed25519
-        .allowlist_type("ed25519_signature")
-        .allowlist_type("ed25519_public_key")
-        .allowlist_function("ed25519_cosi_combine_publickeys")
-        .allowlist_function("ed25519_sign_open")
-        // elligator2
-        .allowlist_function("map_to_curve_elligator2_curve25519")
-        // hmac
-        .allowlist_type("HMAC_SHA256_CTX")
-        .no_copy("HMAC_SHA256_CTX")
-        .allowlist_function("hmac_sha256_Init")
-        .allowlist_function("hmac_sha256_Update")
-        .allowlist_function("hmac_sha256_Final")
-        // sha256
-        .allowlist_var("SHA256_DIGEST_LENGTH")
-        .allowlist_var("SHA256_BLOCK_LENGTH")
-        .allowlist_type("SHA256_CTX")
-        .no_copy("SHA256_CTX")
-        .allowlist_function("sha256_Init")
-        .allowlist_function("sha256_Update")
-        .allowlist_function("sha256_Final")
-        // sha512
-        .allowlist_var("SHA512_DIGEST_LENGTH")
-        .allowlist_var("SHA512_BLOCK_LENGTH")
-        .allowlist_type("SHA512_CTX")
-        .no_copy("SHA512_CTX")
-        .allowlist_function("sha512_Init")
-        .allowlist_function("sha512_Update")
-        .allowlist_function("sha512_Final");
-
-    // Write the bindings to a file in the OUR_DIR.
-    bindings
-        .clang_arg("-Wno-unused-function") // mode_hdr.h has static inline functions unused at parse time
-        .generate()
-        .expect("Unable to generate bindings")
-        .write_to_file(PathBuf::from(out_dir).join("crypto.rs"))
         .unwrap();
 }
