+++ core/embed/rust/src/translations/public_keys.rs
@@ -1,4 +1,4 @@
-use crate::crypto::ed25519;
+use crypto::ed25519;
 
 #[cfg(feature = "dev_keys")]
 pub const PUBLIC_KEYS_DEVEL: [ed25519::PublicKey; 3] = [
