+++ core/embed/sec/Cargo.toml
@@ -10,6 +10,7 @@ xbuild.workspace = true
 
 [dependencies]
 sys.workspace = true
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 
@@ -38,14 +39,14 @@ secure_mode = ["sys/secure_mode"]
 
 backup_ram = []
 bootloader_devel = []
-boot_ucb = ["sys/boot_ucb", "rtl/sphincsplus"]
+boot_ucb = ["sys/boot_ucb", "crypto/sphincsplus"]
 consumption_mask = []
 emulator = ["sys/emulator"]
 hash_processor = []
 hw_revision = []
 iwdg = []
 lockable_bootloader = ["sys/lockable_bootloader"]
-mcu_attestation = ["rtl/mldsa"]
+mcu_attestation = ["crypto/mldsa"]
 nrf_auth = []
 optiga = ["sys/i2c_bus", "time_estimate"]
 optiga_testing = []
