+++ core/embed/sys/Cargo.toml
@@ -9,6 +9,7 @@ color-eyre.workspace = true
 xbuild.workspace = true
 
 [dependencies]
+crypto.workspace = true
 rtl.workspace = true
 models.workspace = true
 
@@ -39,10 +40,11 @@ applets = []
 app_loading = ["applets"]
 block_on_vcp = []
 boot_ucb = []
-emulator = ["rtl/emulator"]
+emulator = ["rtl/emulator", "crypto/emulator"]
 i2c_bus = []
 ipc = []
 lockable_bootloader = []
+production = ["rtl/production", "crypto/production"]
 pvd = []
 rtc = []
 trustzone = []
@@ -53,7 +55,7 @@ dbg_console_system_view = ["dbg_console", "system_view"]
 sdram = []
 sdram_stm32f429i_disc1 = []
 system_view = []
-production = ["rtl/production"]
+
 
 test = [
     "mcu_stm32u5g",
